+++ test/aem-sidekick.test.js
@@ -69,7 +69,7 @@ describe('AEM Sidekick - SSA', () => {
     const { location } = sidekick;
     expect(location.href).to.eq('https://main--aem-boilerplate--adobe.gov-aem.page/');
     expect(sidekickTest.rumStub.called).to.be.true;
-    expect(sidekickTest.rumStub.calledWithMatch('click', {
+    expect(sidekickTest.rumStub.calledWithMatch('top', {
       source: 'sidekick',
     })).to.be.true;
   });